cast float to double

+++ sources-du-projet/src/main/java/com/groupe5/geometry/Point.java
@@ -1,25 +1,25 @@
 package com.groupe5.geometry;
 
 public class Point {
-	private float x;
-	private float y;
-	private float z;
+	private double x;
+	private double y;
+	private double z;
 	
-	public Point(float x, float y, float z) {
+	public Point(double x, double y, double z) {
 		this.x = x;
 		this.y = y;
 		this.z = z;
 	}
 
-	public float getX() {
+	public double getX() {
 		return x;
 	}
 
-	public float getY() {
+	public double getY() {
 		return y;
 	}
 
-	public float getZ() {
+	public double getZ() {
 		return z;
 	}
